Rename avalonia & wpf samples
