+++ samples/CSharpAvalonia/app.manifest
