+++ samples/CSharpAvalonia/build-osx.sh
@@ -15,9 +15,9 @@ RELEASE_DIR="$SCRIPT_DIR/releases"
 PUBLISH_DIR="$SCRIPT_DIR/publish"
 
 echo ""
-echo "Compiling AvaloniaCrossPlat with dotnet..."
-dotnet publish -c Release --self-contained -r linux-x64 -o "$PUBLISH_DIR"
+echo "Compiling VelopackCSharpAvalonia with dotnet..."
+dotnet publish -c Release --self-contained -r osx-x64 -o "$PUBLISH_DIR"
 
 echo ""
 echo "Building Velopack Release v$BUILD_VERSION"
-vpk pack -u AvaloniaCrossPlat -v $BUILD_VERSION -o "$RELEASE_DIR" -p "$PUBLISH_DIR"
+vpk pack -u VelopackCSharpAvalonia -v $BUILD_VERSION -o "$RELEASE_DIR" -p "$PUBLISH_DIR"
