+++ samples/CSharpWpf/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="VeloWpfSample.App"
+﻿<Application x:Class="CSharpWpf.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:VeloWpfSample"
+             xmlns:local="clr-namespace:CSharpWpf"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
