+++ samples/CSharpWpf/AssemblyInfo.cs
