+++ samples/CSharpWpf/CSharpWpf.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
+    <AssemblyName>VelopackCSharpWpf</AssemblyName>
     <TargetFramework>net48</TargetFramework>
     <ImplicitUsings>true</ImplicitUsings>
     <UseWPF>true</UseWPF>
@@ -9,7 +10,7 @@
 
   <PropertyGroup>
     <!-- This overrides the default Program.Main that WPF creates for you, and allows you to add VelopackApp -->
-    <StartupObject>VeloWpfSample.Program</StartupObject>
+    <StartupObject>CSharpWpf.Program</StartupObject>
   </PropertyGroup>
 
   <ItemGroup>
